Изменяет файловую структуру.
