+++ src/components/screens/not-found-screen/not-found-screen.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import Header from "../header/header";
-import Footer from "../footer/footer";
+import Header from '../../header/header';
+import Footer from '../../footer/footer';
+import {Link} from 'react-router-dom';
+import {AppRoutes} from '../../../const';
 
 const NotFoundScreen = () => {
   return (
@@ -14,6 +16,7 @@ const NotFoundScreen = () => {
             <div className="favorites__status-wrapper">
               <b className="favorites__status">Page not found</b>
               <p className="favorites__status-description">Unfortunately the page you are looking for does not exist or there was an error in the link.</p>
+              <p><Link to={AppRoutes.MAIN}>Go to main screen</Link></p>
             </div>
           </section>
         </div>
